+++ lib/src/AtomicModel-UI/utils/ui_colors_palette.dart
@@ -1,37 +1,28 @@
 import 'package:flutter/material.dart';
 
 abstract class UIColorPalette {
-  static const Color colorPrimary = Color(0xFF0D47A1);
-  static const Color colorSecondary = Color(0xFF990000);
-  static const Color colorWhite = Color(0xFFFFFFFF);
 
-  // textBox
-  static const Color trainColorPrimaryTxtBackground = Color(0xFFECEFF1);
-  static const Color trainColorPrimaryTxt= Color(0xFF37474F);
-  static const Color trainColorPrimaryTxtHolder= Color(0xFF90A4AE);
+
+
+  static const Color trainColorPrimaryTxtHolder = Color(0xFF90A4AE);
+
   static const Color secondaryGreyInput = Color(0xFF808fa1);
   static const Color stateRed = Color(0xFFd95a3a);
   static const Color secondaryGreyDisabled = Color(0xFFcfd5dc);
-  static const Color secondaryGrayInput = Color(0xFF808FA1);
+
   static const Color textCharcoal1 = Color(0xFF343c46);
 
-  // texbox search
   static const Color textCharcoal2 = Color(0xFF4c5866);
   static const Color primaryGreyTab = Color(0xFFF0F2F4);
-  static const Color primaryGreyInput = Color(0xFFBFC7D0);
-
-  //button:
 
   static const Color secondarySpaceCadetLight = Color(0xFFcfd0df);
-  static const Color trainColorPrimaryBtn = Color(0xFFFF7043);
-  static const Color primaryWhite = Color(0xFFFFFFFF);
   static const Color trainColorSecondaryBtn = Color(0xFFB0BEC5);
   static const Color primarySpaceCadet = Color(0xFF363853);
-  static const Color primaryK = Color(0xFFCB97FB);
 
-  //menu
   static const Color backgroundContainer = Color(0xffF5F6F9);
   static const Color backgroundColor = Color(0xffdc0a2d);
   static const Color primaryColorLetter = Color(0xffffffff);
   static const Color letterInput = Color(0xff1d1d1d);
+
+  static const Color primaryWhite = Color(0xFFFFFFFF);
 }
